+++ src/plugins/muc-views/templates/muc.js
@@ -1,5 +1,4 @@
 import '../chatarea.js';
-import '../config-form.js';
 import '../destroyed.js';
 import '../disconnected.js';
 import '../heading.js';