+++ src/types/plugins/muc-views/heading.d.ts
@@ -26,7 +26,10 @@ export default class MUCHeading extends CustomElement {
      * @param {Event} ev
      */
     toggleTopic(ev: Event): void;
-    getAndRenderConfigurationForm(): void;
+    /**
+     * @param {Event} ev
+     */
+    showConfigModal(ev: Event): void;
     /**
      * @param {Event} ev
      */