+++ src/types/plugins/muc-views/modals/config.d.ts
@@ -1,0 +1,17 @@
+export default class MUCConfigModal extends BaseModal {
+    renderModal(): import("lit-html").TemplateResult<1>;
+    connectedCallback(): void;
+    getModalTitle(): any;
+    getConfig(): Promise<void>;
+    /**
+     * @param {SubmitEvent} ev
+     */
+    setAvatar(ev: SubmitEvent): Promise<void>;
+    /**
+     * @param {SubmitEvent} ev
+     */
+    submitConfigForm(ev: SubmitEvent): Promise<void>;
+}
+export type VCardData = import('@converse/headless/types/plugins/vcard/api').VCardData;
+import BaseModal from "plugins/modal/modal.js";
+//# sourceMappingURL=config.d.ts.map