+++ src/types/plugins/profile/modals/profile.d.ts
@@ -2,9 +2,16 @@ export default class ProfileModal extends BaseModal {
     tab: string;
     renderModal(): import("lit-html").TemplateResult<1>;
     getModalTitle(): any;
-    setVCard(data: any): Promise<void>;
-    onFormSubmitted(ev: any): void;
+    /**
+     * @param {VCardData} data
+     */
+    setVCard(data: VCardData): Promise<void>;
+    /**
+     * @param {SubmitEvent} ev
+     */
+    onFormSubmitted(ev: SubmitEvent): Promise<void>;
 }
+export type VCardData = import('@converse/headless/types/plugins/vcard/api').VCardData;
 export type XMPPStatus = import("@converse/headless").XMPPStatus;
 import BaseModal from "plugins/modal/modal.js";
 //# sourceMappingURL=profile.d.ts.map